requirements.txt
numpy
pandas
scikit-learn
pyyaml
tqdm
python-dotenv
sentence-transformers
transformers
torch
chromadb
fastapi
uvicorn
pydantic
streamlit
pytest
requests
pypdf
python-multipart
